requirements.txt
requests
python-dotenv
ollama
pyyaml
pdfplumber